# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: M360 API
  version: 1.0.0
  description: |
    M360 SME Lending CRM API for Shariah-compliant Murabaha financing in Saudi Arabia.
    Bilingual Arabic/English with full loan lifecycle management, Saudi government
    integrations (SIMAH, Bayan, Nafath, Yaqeen, Watheq), and role-based access control.

    ## Roles
    - `admin` - Full system access
    - `manager` - Management-level access
    - `loan_officer` - Loan origination and management
    - `credit_analyst` - Credit assessment and committee packages
    - `compliance_officer` - Compliance oversight
    - `collections_officer` - Collections management
    - `data_entry` - Data entry operations

    ## Application Workflow
    Draft -> Submitted -> PreApproved -> DocsCollected -> CreditAssessment -> CommitteeReview -> Approved/Rejected -> Disbursed
  contact:
    name: M360 Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /leads:
    post:
      tags: [Leads]
      summary: Create a lead
      description: Requires admin, manager, loan_officer, or data_entry role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreateRequest'
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Leads]
      summary: List leads
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LeadStatus'
      responses:
        '200':
          description: Paginated list of leads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Lead details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Leads]
      summary: Update a lead
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadUpdateRequest'
      responses:
        '200':
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Leads]
      summary: Delete a lead
      description: Requires admin or manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '204':
          description: Lead deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Products
  # ──────────────────────────────────────────────
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for current user
      parameters:
        - name: unread
          in: query
          description: Filter to unread only
          schema:
            type: string
            enum: ['true', 'false']
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /notifications/count:
    get:
      tags: [Notifications]
      summary: Get unread notification count
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Integrations
  # ──────────────────────────────────────────────
  /integrations/simah/{national_id}:
    get:
      tags: [Integrations]
      summary: Get SIMAH consumer credit report
      description: Retrieves credit score, risk level, active loans, defaults, and total debt for an individual.
      parameters:
        - name: national_id
          in: path
          required: true
          schema:
            type: string
          description: Saudi National ID
      responses:
        '200':
          description: SIMAH credit report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SIMAHCreditReport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/bayan/{cr_number}:
    get:
      tags: [Integrations]
      summary: Get Bayan commercial credit report
      description: Retrieves business credit score, risk level, capital, years in business, and employee count.
      parameters:
        - name: cr_number
          in: path
          required: true
          schema:
            type: string
          description: Commercial Registration number
      responses:
        '200':
          description: Bayan business credit report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BayanCreditReport'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/nafath/{national_id}:
    get:
      tags: [Integrations]
      summary: Verify identity through Nafath
      description: National identity verification via fingerprint, facial recognition, or OTP.
      parameters:
        - name: national_id
          in: path
          required: true
          schema:
            type: string
          description: Saudi National ID
      responses:
        '200':
          description: Nafath identity verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NafathIdentityInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/yaqeen/{national_id}:
    get:
      tags: [Integrations]
      summary: Get person info from Yaqeen
      description: Retrieves enriched personal data including age, gender, marital status, and ID expiry.
      parameters:
        - name: national_id
          in: path
          required: true
          schema:
            type: string
          description: Saudi National ID
      responses:
        '200':
          description: Yaqeen person information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YaqeenIdentityInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /integrations/watheq/{cr_number}:
    get:
      tags: [Integrations]
      summary: Verify CR through Watheq
      description: Verifies Commercial Registration status, entity type, activities, and capital.
      parameters:
        - name: cr_number
          in: path
          required: true
          schema:
            type: string
          description: Commercial Registration number
      responses:
        '200':
          description: Watheq CR verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatheqCRInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Reports
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SearchParam:
      name: search
      in: query
      description: Search term
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (default 20)
      schema:
        type: integer
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
    LeadStatus:
      type: string
      enum: [new, contacted, qualified, unqualified, converted]
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
          nullable: true
        contact_name:
          type: string
        contact_phone:
          type: string
          nullable: true
        contact_email:
          type: string
          nullable: true
        company_name:
          type: string
          nullable: true
        source:
          type: string
        status:
          $ref: '#/components/schemas/LeadStatus'
        estimated_amount:
          type: number
          format: double
          nullable: true
        notes:
          type: string
          nullable: true
        assigned_officer_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LeadCreateRequest:
      type: object
      required: [contact_name, source]
      properties:
        contact_name:
          type: string
        contact_phone:
          type: string
        contact_email:
          type: string
        company_name:
          type: string
        source:
          type: string
        estimated_amount:
          type: number
          format: double
        notes:
          type: string
    LeadUpdateRequest:
      type: object
      properties:
        contact_name:
          type: string
        contact_phone:
          type: string
        contact_email:
          type: string
        company_name:
          type: string
        source:
          type: string
        status:
          $ref: '#/components/schemas/LeadStatus'
        estimated_amount:
          type: number
          format: double
        notes:
          type: string
        organization_id:
          type: string
          format: uuid
    LeadListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        total:
          type: integer

    # ── Product ──
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
          nullable: true
        type:
          type: string
        entity_type:
          type: string
          nullable: true
        entity_id:
          type: string
          format: uuid
          nullable: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Integration responses ──
    SIMAHCreditReport:
      type: object
      properties:
        score:
          type: integer
          description: Credit score (600-870)
        risk_level:
          type: string
          enum: [very_low, low, medium, high]
        active_loans:
          type: integer
        default_count:
          type: integer
        total_debt:
          type: integer
        last_updated:
          type: string
          description: Date string (YYYY-MM-DD)
    BayanCreditReport:
      type: object
      properties:
        score:
          type: integer
          description: Business credit score
        risk_level:
          type: string
          enum: [very_low, low, medium, high]
        active_loans:
          type: integer
        default_count:
          type: integer
        capital:
          type: integer
        years_in_business:
          type: integer
        employee_count:
          type: integer
        last_updated:
          type: string
    NafathIdentityInfo:
      type: object
      properties:
        national_id:
          type: string
        full_name_en:
          type: string
        full_name_ar:
          type: string
        date_of_birth:
          type: string
        nationality:
          type: string
        verified:
          type: boolean
        confidence_level:
          type: number
          format: double
        verification_method:
          type: string
          enum: [fingerprint, facial_recognition, otp]
    YaqeenIdentityInfo:
      type: object
      properties:
        national_id:
          type: string
        full_name_en:
          type: string
        full_name_ar:
          type: string
        date_of_birth:
          type: string
        age:
          type: integer
        gender:
          type: string
        nationality:
          type: string
        id_expiry_date:
          type: string
        marital_status:
          type: string
        verified:
          type: boolean
    WatheqCRInfo:
      type: object
      properties:
        cr_number:
          type: string
        company_name_en:
          type: string
        company_name_ar:
          type: string
        status:
          type: string
          enum: [active, expired, suspended]
        entity_type:
          type: string
        expiry_date:
          type: string
        issue_date:
          type: string
        activities:
          type: string
        capital:
          type: integer
        is_active:
          type: boolean

    # ── Reports ──
...