# backend/internal/lead/handler.go
package lead

import (
	"net/http"
	"strconv"

	"github.com/google/uuid"
	"github.com/labstack/echo/v4"

	"github.com/CamelLabSA/M360/backend/internal/auth"
)

type Handler struct {
	svc *Service
}

func NewHandler(svc *Service) *Handler {
	return &Handler{svc: svc}
}

func (h *Handler) Register(g *echo.Group, authMiddleware echo.MiddlewareFunc) {
	leads := g.Group("/leads", authMiddleware)
	leads.POST("", h.Create, auth.RequireRole(auth.RoleAdmin, auth.RoleManager, auth.RoleLoanOfficer, auth.RoleDataEntry))
	leads.GET("", h.List)
	leads.GET("/:id", h.GetByID)
	leads.PUT("/:id", h.Update)
	leads.DELETE("/:id", h.Delete, auth.RequireRole(auth.RoleAdmin, auth.RoleManager))
}

func (h *Handler) Create(c echo.Context) error {
	var req CreateRequest
	if err := c.Bind(&req); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid request body")
	}
	if err := c.Validate(&req); err != nil {
		return err
	}

	claims := auth.GetClaims(c)
	lead, err := h.svc.Create(c.Request().Context(), req, claims.UserID)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, "failed to create lead")
	}

	return c.JSON(http.StatusCreated, lead)
}

func (h *Handler) GetByID(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}

	lead, err := h.svc.GetByID(c.Request().Context(), id)
	if err != nil {
		return echo.NewHTTPError(http.StatusNotFound, "lead not found")
	}

	return c.JSON(http.StatusOK, lead)
}

func (h *Handler) List(c echo.Context) error {
	limit, _ := strconv.Atoi(c.QueryParam("limit"))
	offset, _ := strconv.Atoi(c.QueryParam("offset"))
	if limit <= 0 {
		limit = 20
	}
	if limit > 100 {
		limit = 100
	}
	if offset < 0 {
		offset = 0
	}

	params := ListParams{
		Search: c.QueryParam("search"),
		Limit:  limit,
		Offset: offset,
	}

	if s := c.QueryParam("status"); s != "" {
		status := Status(s)
		params.Status = &status
	}

	leads, total, err := h.svc.List(c.Request().Context(), params)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, "failed to list leads")
	}

	return c.JSON(http.StatusOK, map[string]interface{}{
		"data":  leads,
		"total": total,
	})
}

func (h *Handler) Update(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}

	var req UpdateRequest
	if err := c.Bind(&req); err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid request body")
	}
	if err := c.Validate(&req); err != nil {
		return err
	}

	lead, err := h.svc.Update(c.Request().Context(), id, req)
	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, "failed to update lead")
	}

	return c.JSON(http.StatusOK, lead)
}

func (h *Handler) Delete(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid id")
	}

	if err := h.svc.Delete(c.Request().Context(), id); err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, "failed to delete lead")
	}

	return c.NoContent(http.StatusNoContent)
}